Simplify benchmarking a lot, avoid useless init/clean as much as possible

+++ nf/testbed/hard/init-machines/middlebox.sh
@@ -2,11 +2,16 @@
 
 echo "[init] Initializing middlebox..."
 
-pushd "$CASE_ROOT"
-  ./vnds/install.sh dpdk-only # this will apt-get update
-popd
+./init-machines/install-dpdk.sh
 
-sudo apt-get install -yqq \
+# CHANGE THIS IF YOU CHANGE THE DEPS BELOW (this allows us to not apt-get update every time)
+DEPS_VERSION='1'
+if [ ! -f '.deps_version' ] || [ "$(cat .deps_version)" != "$DEPS_VERSION" ]; then
+  sudo apt-get update
+  sudo apt-get install -yqq \
     tcpdump git libpcap-dev \
     linux-headers-3.13.0-93 \
     libglib2.0-dev daemon iperf3 netperf tmux
+
+  echo "$DEPS_VERSION" > '.deps_version'
+fi
